MB-52417 Retain covers from parent
As part of the changes for MB-51817/MB-52253, we reset the
covers value from AnnotatedValue during Fetch process, the
intent is to remove _index_key() and _index_condition()
expressions. However, in this defect we have a correlated
subquery, and the outer query is using covering index scan
and the subquery is performing Fetch, and when Fetch is
done we incorrectly removed the covering information for
the parent as well.

The fix is to retain the covering information from the
parent.

Change-Id: I169081c60e152f4231969ced567dd18b9648a1a1
Reviewed-on: https://review.couchbase.org/c/query/+/175639
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>