MB-52254 Remove cached value attachements for early order
As part of Order processing the value for each order expression
is computed once and cached as attachments during the sort
process. For early sort, the cached values are no longer useful
once the sort is done. Add logic to remove cached value
attachments for early order.

Change-Id: Id500e7006a1a06890ede18fa60e3e0cea3a96c1d
Reviewed-on: https://review.couchbase.org/c/query/+/175676
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/175684
Well-Formed: Restriction Checker