MB-52176 Schedule UPDATE STATISTICS after CREATE INDEX and BUILD INDEX
To assist in using CBO for potential cost for queries in Elixir
we now automatically post a task to the scheduler to perform
"UPDATE STATISTICS FOR INDEX" when a CREATE INDEX statement
finishes (if not deferred), or when a BUILD INDEX statement
finishes (for deferred). There is no guarantee that the
UPDATE STATISTICS statement will finish without error and
optimizer statistics will be available, but that's the best
we can do I guess. (The error, if any, from UPDATE STATISTICS
statement will be kept in the system:tasks_cache).

Change-Id: I603e6c004500f038848c50f34f00b7534fa25057
Reviewed-on: https://review.couchbase.org/c/query/+/175824
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>