MB-52176 Run unit tests with CBO off
Since we now do UPDATE STATISTICS automatically with
CREATE INDEX statement, and CBO is turned on by default,
some unit tests are failing (different plan chosen) due
to CBO kicking in. This is not always deterministic since
there is a slight delay between index create and the
corresponding UPDATE STATISTICS statement, thus it is not
guaranteed that UPDATE STATISTICS has been run when
subsequent statements are being executed. To avoid this
uncertainty, turn off CBO when running unit tests.
Individual tests can still turn on CBO by adding a new
query parameter in the test case.
Also need to check on ORDERED hint regardless of whether
CBO is on or off.

Change-Id: I36bcc6363bef2b899123a55a956673bc75d4ea92
Reviewed-on: https://review.couchbase.org/c/query/+/175917
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>