MB-52503 Avoid early order if ORDER BY expression depends on LET variable
The problematic query has ORDER BY clause that uses a LET variable,
and the LET binding has a correlated subquery. We should not be
using early order in this case since the LET variable is not
evaluated until later on.

The fix is to detect situations where ORDER BY expression depends
on a LET variable, and disable early order in such cases.

Also avoid early index filter if the filter contains a subquery
(since the filter is evaulated twice).

Change-Id: I33bba0aa2943d40b6c6a0b55eb7fa6f6344c6159
Reviewed-on: https://review.couchbase.org/c/query/+/176069
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/176070
Well-Formed: Restriction Checker