MB-52509 Handle query parameters for LIMIT/OFFSET when considering early order
When a query uses query parameters (named/positional parameters) for
LIMIT and/or OFFSET specification, currently early order is not
considered, since the cost for the ORDER operator cannot be easily
computed (without the actual number for LIMIT/OFFSET).

Detect and allow query parameters to be used for LIMIT/OFFSET
specification, and for costing purpose, if LIMIT/OFFSET is not
available, assume half of the documents from the index scan will
be returned from the ORDER operator, for costing purpose. This
way we still allow early order to be used.

Change-Id: I4c59fde6368fd2ea766a0b49d71589bbb779ef8e
Reviewed-on: https://review.couchbase.org/c/query/+/176198
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>