MB-52689 Save/restore of advisor structure for subquery
The main issue with this defect is that when we ADVISE for
subqueries, there are a couple of additional structures
in the builder that requires save/restore operation. This
is in addition to the queryInfos map that is currently
being saved/restored.

Another issue is when we run ADVISE on subquery, if
there is an error from the planner, do not simply return
the error (e.g. no INDEX available). Let the logic
continues such that we can advise indexes later.

Change-Id: I5e8507a8b227106c51bc252361169d1fd3fd35bf
Reviewed-on: https://review.couchbase.org/c/query/+/176692
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>