MB-52873 Use Greatest/Least expression when query parameter is used
When multiple predicates can be used to generate index span on
the same index key, we "constrain" the index span by trying to
use information from both index spans. When a query parameter
is involved, since we cannot choose based on actual value,
previous code just use the one with the pamater to be on the
safe side. This in some cases can result in a span that's much
larger than necessary, and poor performance.

The fix is to detect such situations and use Greatest() function
on the low bound and Least() function on the high bound, and
thus let runtime choose the one when parameter value is
available.

Change-Id: I3cda76fc0d7249f33f05af501a47ab02a5b8e45a
Reviewed-on: https://review.couchbase.org/c/query/+/177244
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>