MB-52899 Avoid generating empty index projection for missing leading key index
With a missing leading key index, it is possible to generate
an empty index projection if this index is used more like a
primary index, and none of the index keys are referenced in
the query.

The fix is to detect "empty" index projection and include
primary key in the index projection.

Change-Id: If970b870ac192e224d20325846e518b2c4cf3ca0
Reviewed-on: https://review.couchbase.org/c/query/+/177315
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
(cherry picked from commit 9950ec7414ea078bcf293dc47498062a065ba648)
Reviewed-on: https://review.couchbase.org/c/query/+/177058
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker