MB-52921 Exclude sequential scan index in existing index infos
When collecting existing index infos, for the sequential scan
index, the fix for MB-52917 results in a nil pointer being
returned. The issue with this defect is that the nil pointer
is being added as an existing index, causing nil pointer
de-reference later on.

The fix is to check for nil pointer before adding the index
into existing index infos.

Change-Id: I826088db6161ab337a946be9b980f7005cb59203
Reviewed-on: https://review.couchbase.org/c/query/+/177383
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>