MB-51977 No need to drop from _system scope if bucket is being dropped
In serverless mode, CBO stats are stored under _system scope
for each bucket. If a bucket is being dropped, we have logic
to remove CBO stats for the collections in that bucket, which
includes dropping from the _system scope. This is not
necessary anymore since the whole bucket is being dropped.

Check and ignore the step for dropping from _system scope
in serverless mode, if the bucket is being dropped.

Change-Id: I5e4186d332e5fd1d7eba9d2ea8ccd79ab86c7df9
Reviewed-on: https://review.couchbase.org/c/query/+/177870
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>