MB-51961 Allow hash join to be used in RBO if nested-loop join is not available
Currently in RBO, hash join is only considered when a USE HASH hint
is specified. For nested-loop join, which currently requires an
appropriate secondary index on the inner of the join, an error will
be returned if such a secondary index does not exist.

To improve developer experience, especially someone who is not
familiar with Couchbase and just want to try some queries, giving
an error on a join query due to lack of a secondary index is not
ideal. We now include consideration of hash join for RBO even in
the absence of a USE HASH hint. A hash join is used in RBO only
when nested-loop join is not feasible, e.g., lack of an appropriate
secondary index (with CBO this will be a cost-based decision).

Note this is only applicable to EE since hash join is not enabled
in CE.

Change-Id: I4e1574e2292b75989b04472b6566c98bc5a9af88
Reviewed-on: https://review.couchbase.org/c/query/+/177907
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>