MB-51961 Allow primary index to be used on inner of nested-loop join
Currently primary index is not allowed to be used on the inner of a
nested-loop join. This restriction was put in place to avoid a "bad"
query plan. However, for a new developer who may not be as familiar
with Couchbase and is just playing around with simple queries,
getting an error (lack of secondary index) may not be desired.

Previous checkin allows hash join to be considered if nested-loop
join is not feasible (e.g. lack of appropriate secondary index),
which should greatly reduce the occurrence of errors from join
queries. However hash join is only feasible when equi-join is being
performed.

With this checkin, we'll further reduce the occurrence of error
from join queries by allowing primary index (or kv range scan) to
be used on the inner of nested-loop join.

The preference of join methods is as follows:

With CBO, both hash join and nested-loop join are considered and
a join method is chosen by cost.

With RBO, nested-loop join is the default join method, a hash join
is only considered when a nested-loop join is not available; or
when a USE_HASH hint is specified. (Note that even without a join
hint hash join is now considered as well).

If neither nested-loop join nor hash join is feasible, then as a
last resort we now consider using either a primary index or a
kv range scan under a nested-loop join. This is the case where a
join is done on non-equality conditions, and appropriate secondary
index is not available.

To avoid a "bad" plan and still allow simple queries to go through,
an arbitrary restriction is put in place, such that when a primary
index (or kv range scan) is being used under a nested-loop join,
it can at most produce 1000 documents. If this limit is exceeded,
we either return the same error as before (lack of appropriate
secondary index on inner), if this can be detected at compile time,
or if this is detected at runtime, a new error (code 5420) will be
returned.

As a further performance optimization, when a primary index is used
under the inner of nested-loop join, we "cache" the result of such
primary index scan, such that subsequent join operations can be
performed on the cached result, instead of repeating the primary
index scan. Such caching is either done on IndexScan3 (in case of
covering primary index scan) or Fetch (in case of non-covering
primary index scan). When such cache is used we also set
maxparallelism to 1 such that a single cache is maintained.

Change-Id: I602d03663ce6fddfeced926719931c562b0ffb66
Reviewed-on: https://review.couchbase.org/c/query/+/178374
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>