MB-53298 Do not return error for MISSING value in join processing
Previous code for join assumes that the inner of a join should have
a valid value, since the join is being processed after a potential
match has been found. However, in this bug, we are doing the
equivalent of a cartesian join, and an explicit MISSING value is
used on the inner side of a join. This triggers a sanity check
failure.

The fix is to allow such MISSING value to proceed with join
processing. Also optimizer hints processing is fixed when joining
on expression/subquery terms.

Change-Id: I2ce8ee1759dc1547451241cfa6f19102d089639b
Reviewed-on: https://review.couchbase.org/c/query/+/178581
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>