MB-53353 Adjust unbounded folllowing window range
Incremental aggregates work by first aggregating all items in the window
then for the next row shifting the window on by one and removing the row
that has fallen out of the window and adding the row that has entered.
With unbounded following, there should never be anything entering only
gradual depletion as the advancing window drops rows.

With the upper index set to length-1, the final item is aggregated
repeatedly.  For example:

SELECT
x.a AS a,
SUM(x.a) OVER (PARTITION BY x.c ORDER BY x.a ROWS BETWEEN CURRENT ROW AND unbounded FOLLOWING) AS aSum
FROM [
{"a":0,"c":'one'},
{"a":1,"c":'one'},
{"a":4,"c":'one'},
{"a":7,"c":'one'}
] AS x
ORDER BY a ASC;

yields (before fix):

    "results": [
    {
        "a": 0,
        "aSum": 12     <= 0 + 1 + 4 + 7
    },
    {
        "a": 1,
        "aSum": 19     <= 12 - 0 + 7
    },
    {
        "a": 4,
        "aSum": 25     <= 19 - 1 + 7
    },
    {
        "a": 7,        <= 25 - 4 + 7
        "aSum": 28
    }

It can be further highlighted by changing a:7 to a:8 - the figures are
then 13, 21, 28 & 32.

The fix sets the final index beyond the last row so it is only included
once.

Change-Id: I3b77cd51e151275a04a0f42cf6138b4412c91efe
Reviewed-on: https://review.couchbase.org/c/query/+/178799
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>