MB-53377 Proper set up of join keys when converting from ANSI join to lookup join
If an ANSI join is being performed on document key: meta(rhs).id = <expr>,
then in case there is no possible index to construct a join, we internally
convert the join to lookup join. The issue with this bug is that in case
an error is returned, we did not properly return the constructed join key
expression for the lookup join. (Previous code handled the most common
error in this case, which is E_NO_ANSI_JOIN, and is ignored, however in
this bug a different error is returned).

THe fix is to return the proper primaryJoinKeys even in case of error.

Change-Id: Iba5c7fd27a3e8f9e1469446ba78825f79097a9da
Reviewed-on: https://review.couchbase.org/c/query/+/178823
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/178824
Well-Formed: Restriction Checker