MB-53956 Fix value tracking for RAW projections
When the expression evaluated for a RAW projection is "self", the input
value item is returned as the result.  This is then set as the
projection and is cleaned up with the forwarded value.  If it isn't
"self", the the result of the expression is a new value which means
item should be tidied up here.

Change-Id: Id0c5f02f53d39b9b83ee646966d8c14cabc5b774
Reviewed-on: https://review.couchbase.org/c/query/+/178933
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
(cherry picked from commit 6502af299a68cce3152828d72751d6a1b02a1fcd)
Reviewed-on: https://review.couchbase.org/c/query/+/182636
Well-Formed: Restriction Checker
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>