MB-53444 Avoid invalid numbers when calculating group cost
Previous calculations of group cost depends on the input cardinality
and uses default distinct fractions if statistics is not available
on grouping expression. As a result valid numbers are always returned
as group cardinality. Enhancements done in 7.2 also takes into
account document counts of the keyspaces involved in the grouping
expressions. The issue is that if statistics is not available then
document count is -1, and can result in invalid calculations.

The fix is to detect such situations and "bail out" of cost
calculations. This is consistent with other cost calculations where
we return -1 for cost/cardinality etc when statistics is not
available.

Change-Id: Idcf4b6e4b34155b79d23128fca036f609bcad3c5
Reviewed-on: https://review.couchbase.org/c/query/+/179039
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>