MB-53446 Honor USE NL hint for kv range scan
Previous change for MB-51961 allows kv range scan to be
used on the inner of nested-loop join. However, if hash
join is available, it is preferred over nested-loop join
using kv range scan.

When a USE NL hint is specified, we should try to honor
the hint and use nested-loop join, even when kv range
scan is used on the inner.

Modify the logic such that we delay returning of hash
join in case USE NL is specified, until after
consideration of nested-loop join using kv range scan
on inner.

Change-Id: I596787f066f7a142bbec5fa23f5a2a57086a444a
Reviewed-on: https://review.couchbase.org/c/query/+/179041
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>