MB-51977 select from system:dictionary in serverless mode
In serverless mode, gather a list of buckets that user has access to,
and for each bucket, get the necessary information from the _query
collection in the _system scope. Also need to conver the document key
which has "cbo::" prefix and includes only scope and collection name,
this is converted to the "regular" qualified name format.

Change-Id: Idbbb70e1351987430c2dfb6ea8b0090d17efbf99
Reviewed-on: https://review.couchbase.org/c/query/+/179418
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>