MB-53565 Do not use join predicates as index span if under hash join
When generating an index span for an index scan under hash join, we
should not use join predicates since we cannot evaluate expressions
from the other keyspace yet. This behavior is correct in 6.5 and 6.6,
however, in 7.0 and above, with the fix for MB-30813, in order to
allow correlation variables to be used in an index span, we added
an additional check for whether the query block is correlated. The
issue with that is the check does not properly distinguish between
a correlation variable and a reference to a different keyspace in
the same query block.

The fix is to add a check in case of correlated subuqery, to see
whether the expression references any other keyspaces in the same
query block.

Change-Id: I727360eff8b0014129fe44e1210f1041ed34aba0
Reviewed-on: https://review.couchbase.org/c/query/+/179432
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>