MB-53565 Proper formalization for ExpressionTerm and SubqueryTerm under ANSI join
When formalizing an ExpressionTerm or a SubqueryTerm, currently a new
formalizer is always created and returned. A new formalizer is necessary
to check whether there is any correlation in the expression for ExressionTerm
or the subuqery for SubqueryTerm. However, if this ExpressionTerm or SubqueryTerm
is under an ANSI join/nest operation, we need to use (and return) the parent's
formalizer instead of the newly created formalizer. This logic already exists
for KeyspaceTerm, however it is not done for ExpressionTerm and SubqueryTerm.
In the problematic query two ExpressionTerms are used in a join operation, and
after formalizing both, the formalizer only contains the second term as
"immediate" term, and thus the join expression with the first term is incorrectly
detected as "correlated".

The fix is to check and use the parent's formalizer in case of ANSI join/nest,
similar to what's already done for KeyspaceTerm.

Change-Id: Ib5b770c74c2288b75f33e59402300378d85d18f1
Reviewed-on: https://review.couchbase.org/c/query/+/179438
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>