MB-53565 Do not consider primary index for index scan inside correlated subquery
We don't allow primary index scans for a keyspace inside a correlated
subquery (for performance reasons). Current code would add primary
index to the list of indexes to be considered, and it is possible
that when CBO is on we could prefer a primary index due to lower
cost. Later on we detect primary index and return an error, even
though a usable secondary index exists.

The fix is to exclude a primary index for a keyspace inside a
correlated subquery.

Change-Id: Ibd5544c6df96155b5df219072e20f69f29ee444f
Reviewed-on: https://review.couchbase.org/c/query/+/179440
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>