MB-53601 Do not allow hash join under the inner of nested-loop join
Hash join should not be under the inner of a nested-loop join
since the inner side of nested-loop join requires the "reopen"
logic that is not implemented for hash join.

The only chance this could happen is when a SubqueryTerm is
used on the inner side of a join, and the join is using
nested-loop join for whatever reason (e.g. in case of
correlation in which case hash join is not feasible). Then
inside the SubqueryTerm a join could be using hash join.

The fix is to mark the builder when the inner of a
nested-loop join is being planned, and avoid consideration
of hash join in such cases.

Change-Id: I454245e75cff01484cc198a72965cd1135ff1e80
Reviewed-on: https://review.couchbase.org/c/query/+/179599
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>