MB-53601 Use ExpressionScan for SubqueryTerm if under inner of nested-loop join
When a FROM clause subquery (SubqueryTerm) is used under the inner
side of a nested-loop join, use an ExpressionScan on top of the
subquery instead of embedding the subquery plan under the inner
of nested-loop join. This way, the "reopen" logic only needs to
work properly for ExpressionScan and we don't need to implement
"reopen" for all operators.

An added benefit is that if the subquery is not correlated, then
the result of the subquery will be cached and thus not executed
repeatedly under nested-loop join (which will be the case if the
subquery plan is embedded in the parent query plan).

If the subquery is correlated, then the ExpressionScan will
automatically re-execute the subquery.

A few additional things that need to be taken care of:

1. From clause subqueries do not appear in the "~subqueries"
   section of the explain since it is normally inlined. If we
   now put an ExpressionScan on top of the subquery, we need
   to add the subquery and its plan to "~subqueries" section.

2. If there are any optimizer hints in the SubqueryTerm,
   previously these hints are gathered into the parent
   query's optimizer hints and displayed together with
   the parent query. Now if we have a separate section
   under "~subqueries" for this subquery, its optimizer
   hints will be displayed in that section, and thus we
   also need logic to remove these hints from the parent
   query's optimizer hints.

3. The ExpressionScan itself can cache result of the
   expression evaluation if not correlated. In case we
   have a subquery as the expression, the subuqery
   evaluation also caches results, thus we don't need to
   cache the results again at ExpressionScan. Added logic
   to avoid caching result at ExpressionScan in this case.

4. With this fix, the previous restriction of not allowing
   hash join under inner of nested-loop join is obsolete,
   reverse part of previous fix.

Change-Id: I34f584fda25c7609bea8b6e6fda6875f3340cdaf
Reviewed-on: https://review.couchbase.org/c/query/+/179754
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>