MB-22424 Add sequences to scopes
Sequences are a scope-level artefact only.

Privileges: manage & use, assigned at the scope level.

Storage is in the system storage location - _system._query collection.
If this collection doesn't exist for a bucket, sequences are not enabled
for it (migrating instances can't start using sequences in a bucket
until this has been created).

metakv is only used to synchronise nodes.

Backup/restore of sequences is part of the system keyspace
backup/restore; there is no special processing for sequences.

Numbers are stored as strings so we can have the full 64-bit signed
integer range without potential impact (particularly since the default
range is Int64Min to Int64Max).

See the MB for overview document link.

https://docs.google.com/document/d/1UMv0GsLIrGSgNu4uW_5Yqc4QgiUPF-nrs3gGgWa9zYU/edit?usp=sharing

Change-Id: Ifb76cd5033062a1b25fd14daf885ac5e2e9bd2d9
Reviewed-on: https://review.couchbase.org/c/query/+/179864
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>