MB-53600 Skip SubqueryTerm plan when checking build side aliases
As part of join filter (bloom filter) consideration, there is
a process of removing some aliases from consideration, e.g.,
the inner side of a nested-loop join/nest. We walk the existing
plan looking for nested-loop join/nest for this purpose. The
issue with this bug is that this process descends into the
(embedded) query plan for a SubqueryTerm, which contains aliases
that's only visible inside the subquery, and triggers a sanity
failure.

We already have logic to skip the query plan for a SubqueryTerm
in several other functionality for bloom filter, just need to
also include this logic when walking the plan and looking for
aliases to skip.

Change-Id: I54ec53669a0538bbea457fba5aed68f3b1babd23
Reviewed-on: https://review.couchbase.org/c/query/+/179925
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>