MB-53774 Properly build hash table for IN-list evaluation if first value is NULL
For IN-list evaluation if the length of the IN-list is > threshold (16)
we use a hash table for evaluation. The issue with this defect is that
if the first value to be checked happens to be a NULL value, then
the logic for building the hash table misbehaves and the hash table is
not properly built, resulting in incorrect evaluation.

The fix is to ensure that the hash table is properly built in case the
first value to be checked is a NULL value. Also if the value to be
checked is a NULL value, there is no need to probe the hash table.

Change-Id: Ie2dd594c3e51fa0683b4fbd778f943ee602f17e0
Reviewed-on: https://review.couchbase.org/c/query/+/180274
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>