MB-53770 Enchance TO_NUMBER.
This allows an optional second argument to TO_NUMBER which is a list of
characters to be stripped from the first argument string (if it isn't a
string the second argument is ignored).

In addition, if the second argument is passed, all space characters are
stripped (throughout).

If a solitary comma is detected as the decimal separator, it is
permitted.

If the second argument isn't passed, it operates as before (without any
character removal or conversion).

This allows all of these:

TO_NUMBER("1 7 5 9 1 . 4 6","")
TO_NUMBER("£17 591.46","£")
TO_NUMBER(" $ 17,591.46","$,")
TO_NUMBER("AU$17,591.46 ","AU$,")
TO_NUMBER("€   17.591,46 ","€.")

to correctly parse as: 17591.46

Change-Id: Iffcfe72458bab4ea6ca9f9ebe4703db3066f5cee
Reviewed-on: https://review.couchbase.org/c/query/+/180286
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>