MB-53774 Handle NULL value separately
When the left-hand side of an IN expression has NULL value,
current logic still either loop through the IN-list or attempt
to check the hash table, except the only thing it does is to
set null = true, thus the only thing it achieves is to
distinguish whether the array on the right-hand side of IN
expression is an empty array or not.

Pull the check for NULL value from the left-hand side of IN
expression to earlier in the function. This should simplify
the logic for the rest of the function.

Change-Id: Ie66708e5285ec1fb7cd058ecbbe061f16c716071
Reviewed-on: https://review.couchbase.org/c/query/+/180312
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>