MB-53793 Handle OR predicate better
For OR predicates we go through two passes, one when we
consider the entire OR clause, and one when we use individual
subterms of the original OR predicate. Mark the case where we
use the entire OR clause and in such case allow using
individual filters to sargIndexes.

Change-Id: Idb5a07b6d38a1e188428ad201b0c5270a0ba83db
Reviewed-on: https://review.couchbase.org/c/query/+/180392
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>