MB-53815 remove system dependency from cbq
cbq uses godbc which uses primitives.
having primitives reference tenant means that cbq would have to link system and sigar...

Solutions:
godbc not to use primitives
primitives not to use tenant
tenant and memory not to use system (this is in fact doable, as the only reason to do that is to give available memory to context)

for now, we just give primitives an indirect runtime reference to tenant.

Change-Id: If3aab73ed89b0b93b3b1c9591115bf9db0696e1c
Reviewed-on: https://review.couchbase.org/c/query/+/180502
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>