MB-53954 Allow sort to spill for UPDATE STATISTICS
UPDATE STATISTICS performs sort operations on each term
for which a histogram is to be constructed. Allow these
sorts to spill to disk in order to minimize the number
of batches and thus reduce access to kv engine.

The logic for determining when to spill is similar to
what is used for the ORDER operator with some adaptation.

Change-Id: I49e053c48a2537dc5675fdfa70e8d909870ec3e6
Reviewed-on: https://review.couchbase.org/c/query/+/180746
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>