MB-53905 Ensure entire collection can be sampled
When a collection's size is smaller than the requested sample size we
restrict the sample size to the document count.  For a random scan, we
normally divide the sample size by the number of v-buckets for the
per-v-bucket size; this can lead to under sampling owing to uneven
distribution amongst v-buckets.  This changes the approach when the
entire collection should be sampled to using the maximum possible sample
size for every v-bucket ensuring complete sampling.

Change-Id: I4eeffe10fa5741cf78606e4d075eae51e3c37864
Reviewed-on: https://review.couchbase.org/c/query/+/180812
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>