MB-54052 Retry conn & create for sequential scans
When there is a problem connecting to the KV reflected in an existing
connection _at scan creation time_ or in getting a v-bucket connection,
we'll now retry the individual scan a number of times with a regular
delay between attemtps.  Failures with existing connections during scan
processing are reported immediately and scans are not retried.

Change-Id: Ic4fce4f6172e57af86caff5c7733be7ec4695f6f
Reviewed-on: https://review.couchbase.org/c/query/+/180998
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>