MB-54406 Add resident ratio to keyspace stats in dictionary
When gathering basic keyspace stats, e.g. document count,
average document size, etc, also now gather memory usage
from memcached, such that we can calculate a resident
ratio for a collection. This will be used later in I/O
cost calculation when a FETCH is performed.

There is some issue with compressed size vs uncompressed
size so the numbers may not be as accurate currently,
hopefully this will be fixed.

Note this is only available for a collection, for bucket,
or other "fake" keyspaces such as mock, file, system,
virtual, use -1 (invalid) to represent resident ratio.

Change-Id: I032b87aed23810f7003cf1d978f40f0ed8787489
Reviewed-on: https://review.couchbase.org/c/query/+/182280
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>